+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ### Resources
 * **New Resource:** `identitycloud_variable` ([#7](https://github.com/pingidentity/terraform-provider-pingfederate/pull/7))
 * **New Resource:** `identitycloud_custom_domains` ([#8](https://github.com/pingidentity/terraform-provider-pingfederate/pull/8))
+* **New Resource:** `identitycloud_certificate` ([#9](https://github.com/pingidentity/terraform-provider-pingfederate/pull/9))